requirements.txt
suds>=0.4
leancloud-sdk==0.10.0